requirements.txt
jira==2.0.0
Django>=2.0,<3.0
djangorestframework==3.9.0
python-dotenv